avoid BDD return type, use ret_ptr instead to also work on ARM
This makes HasCacBDD work on aarch64 including ARM and Apple M1.

Big thanks to Melissa aka shadysciencefeline for help!

Previously we relied on calling conventions specific to amd64:

> "When an oversized struct return is needed, another pointer to a
> caller-provided space is prepended as the first argument, shifting
> all other arguments to the right by one place ..."
https://en.wikipedia.org/wiki/X86_calling_conventions#System_V_AMD64_ABI

Hence the difference between this in C:

  BDD BDD_Operator_Not(const BDD* this_ptr);

and this in Haskell:

  foreign import ccall unsafe "BDDNodeC.h BDD_Operator_Not" bdd_Operator_Not :: Ptr CacBDD -> Ptr CacBDD -> IO (Ptr CacBDD)

But that calling convention is not working on arm, and resulted in the mysterious `0x8`:

> x8 (XR): Indirect return value address.
https://en.wikipedia.org/wiki/Calling_convention#ARM_(A64)

To be compatible with both calling conventions we now do what amd64 does manually, with this in C:

  void BDD_Operator_Not(BDD* ret_ptr, const BDD* this_ptr);

and this in Haskell:

  foreign import ccall unsafe "BDDNodeC.h BDD_Operator_Not" bdd_Operator_Not :: Ptr CacBDD -> Ptr CacBDD -> IO ()

(In the actual code we still abbreviate using `UnaryOp`.)