do bdd_new in C, not in Haskell
To work on aarch64 too we used `b <- bdd_new` to set the return value
ret_ptr. Preliminary benchmarking suggests decreased performance with
this, so here we try to improve it again by doing `new BDD()` on the C
side and removing `bdd_new` from the Haskell side.

So far it does not seem to help, though.